Drop duplicate TypeScript smoke files and package LICENSE copies.
