+++ tenuo-ts/.gitignore
@@ -1,6 +1,8 @@
 node_modules/
 dist/
 *.tsbuildinfo
+packages/core/LICENSE
+packages/mcp/LICENSE
 packages/core/src/generated/README.md
 packages/core/src/generated/.gitignore
 packages/core/src/generated/.npmignore
