+++ tenuo-ts/packages/core/test/mcp.test.ts
@@ -473,3 +473,18 @@ describe("tenuo.mcp", () => {
     ).resolves.toMatchObject({ path: "/data/q3.pdf" });
   });
 });
+
+describe("mcp smoke", () => {
+  it("runs the quarterly-close scenario: narrow, deny, approval, and forged args", async () => {
+    const { runMcpSmoke } = await import("../examples/mcp/scenario.ts");
+    const result = await runMcpSmoke();
+    expect(result.listed).toEqual(["q3-customers.md", "q3-revenue.md"]);
+    expect(result.emailed).toMatchObject({ to: "finance@acme.com" });
+    expect(result.denials.map((denial) => denial.code)).toEqual([
+      "TENUO_CONSTRAINT_VIOLATION",
+      "TENUO_CONSTRAINT_VIOLATION",
+      "TENUO_APPROVAL_REQUIRED",
+      "TENUO_INVALID_POP",
+    ]);
+  });
+});
